+++ image-loader/app/ImageLoaderComponents.scala
@@ -3,7 +3,7 @@ import com.gu.mediaservice.lib.play.GridComponents
 import controllers.ImageLoaderController
 import lib._
 import lib.storage.ImageLoaderStore
-import model.{ImageUploadOps, ImageUploadProjector}
+import model.{Uploader, Projector}
 import play.api.ApplicationLoader.Context
 import router.Routes
 
@@ -17,11 +17,12 @@ class ImageLoaderComponents(context: Context) extends GridComponents(context) {
 
   val notifications = new Notifications(config)
   val downloader = new Downloader()
-  val imageUploadOps = new ImageUploadOps(store, config, imageOperations)
+  val uploader = new Uploader(store, config, imageOperations, notifications)
 
-  val imageUploadProjector = ImageUploadProjector(config, imageOperations)
+  val projector = Projector(config, imageOperations)
 
-  val controller = new ImageLoaderController(auth, downloader, store, notifications, config, imageUploadOps, imageUploadProjector, controllerComponents, wsClient)
+  val controller = new ImageLoaderController(
+    auth, downloader, store, notifications, config, uploader, projector, controllerComponents, wsClient)
 
   override lazy val router = new Routes(httpErrorHandler, controller, management)
 }
