+++ image-loader/app/model/Projector.scala
@@ -1,27 +1,31 @@
 package model
 
-import java.io.File
+import java.io.{File, FileOutputStream}
 
-import com.amazonaws.services.s3.model.ObjectMetadata
+import com.amazonaws.services.s3.AmazonS3
+import com.amazonaws.services.s3.model.{ObjectMetadata, S3Object}
 import com.gu.mediaservice.lib.ImageIngestOperations
 import com.gu.mediaservice.lib.aws.S3Ops
 import com.gu.mediaservice.lib.imaging.ImageOperations
 import com.gu.mediaservice.lib.logging.RequestLoggingContext
 import com.gu.mediaservice.lib.net.URI
 import com.gu.mediaservice.model.{Image, UploadInfo}
-import lib.imaging.MimeTypeDetection
+import lib.imaging.{MimeTypeDetection, NoSuchImageExistsInS3}
 import lib.{DigestedFile, ImageLoaderConfig}
+import org.apache.tika.io.IOUtils
 import org.joda.time.{DateTime, DateTimeZone}
+import play.api.Logger
 
 import scala.collection.JavaConverters._
-import scala.concurrent.{ExecutionContext, Future}
+import scala.concurrent.duration.Duration
+import scala.concurrent.{Await, ExecutionContext, Future}
 
-object ImageUploadProjector {
+object Projector {
 
-  import ImageUploadOps.toImageUploadOpsCfg
+  import Uploader.toImageUploadOpsCfg
 
-  def apply(config: ImageLoaderConfig, imageOps: ImageOperations)(implicit ec: ExecutionContext): ImageUploadProjector
-  = new ImageUploadProjector(toImageUploadOpsCfg(config), imageOps)(ec)
+  def apply(config: ImageLoaderConfig, imageOps: ImageOperations)(implicit ec: ExecutionContext): Projector
+  = new Projector(toImageUploadOpsCfg(config), S3Ops.buildS3Client(config), imageOps)(ec)
 }
 
 case class S3FileExtractedMetadata(
@@ -54,10 +58,38 @@ object S3FileExtractedMetadata {
   }
 }
 
-class ImageUploadProjector(config: ImageUploadOpsCfg, imageOps: ImageOperations)
-                          (implicit val ec: ExecutionContext) {
+class Projector(config: ImageUploadOpsCfg,
+                s3: AmazonS3,
+                imageOps: ImageOperations)
+               (implicit val ec: ExecutionContext) {
 
   private val imageUploadProjectionOps = new ImageUploadProjectionOps(config, imageOps)
+
+  def projectS3ImageById(imageUploadProjector: Projector, imageId: String, requestLoggingContext: RequestLoggingContext, tempFile: File)(implicit ex: ExecutionContext): Future[Option[Image]] = {
+    Logger.info(s"Projecting image: $imageId")(requestLoggingContext.toMarker())
+    Future {
+      import ImageIngestOperations.fileKeyFromId
+      val s3Key = fileKeyFromId(imageId)
+
+      if (!s3.doesObjectExist(config.originalFileBucket, s3Key))
+        throw new NoSuchImageExistsInS3(config.originalFileBucket, s3Key)
+
+      Logger.info(s"object exists, getting s3 object at s3://${config.originalFileBucket}/$s3Key to perform Image projection")(requestLoggingContext.toMarker())
+
+      val s3Source = s3.getObject(config.originalFileBucket, s3Key)
+      val digestedFile = getSrcFileDigestForProjection(s3Source, imageId, requestLoggingContext, tempFile)
+      val extractedS3Meta = S3FileExtractedMetadata(s3Source.getObjectMetadata)
+
+      val finalImageFuture = imageUploadProjector.projectImage(digestedFile, extractedS3Meta, requestLoggingContext)
+      val finalImage = Await.result(finalImageFuture, Duration.Inf)
+      Some(finalImage)
+    }
+  }
+
+  private def getSrcFileDigestForProjection(s3Src: S3Object, imageId: String, requestLoggingContext: RequestLoggingContext, tempFile: File) = {
+    IOUtils.copy(s3Src.getObjectContent, new FileOutputStream(tempFile))
+    DigestedFile(tempFile, imageId)
+  }
 
   def projectImage(srcFileDigest: DigestedFile, extractedS3Meta: S3FileExtractedMetadata, requestLoggingContext: RequestLoggingContext): Future[Image] = {
     import extractedS3Meta._
@@ -90,7 +122,7 @@ class ImageUploadProjector(config: ImageUploadOpsCfg, imageOps: ImageOperations)
 class ImageUploadProjectionOps(config: ImageUploadOpsCfg,
                                imageOps: ImageOperations)(implicit val ec: ExecutionContext) {
 
-  import ImageUploadOps.{fromUploadRequestShared, toMetaMap}
+  import Uploader.{fromUploadRequestShared, toMetaMap}
 
   private val dependenciesWithProjectionsOnly = ImageUploadOpsDependencies(config, imageOps,
     projectOriginalFileAsS3Model, projectThumbnailFileAsS3Model, projectOptimisedPNGFileAsS3Model)
